requirements.txt
python-dotenv==1.0.0
nest-asyncio==1.5.8
pydantic-ai==0.1.2 